+++ src/main/java/com/dsf/escalade/repository/business/SecteurRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SecteurRepository extends JpaRepository <Secteur, Integer>  {
+public interface SecteurRepository extends JpaRepository <Secteur, Long>  {
 }
