+++ src/main/java/com/dsf/escalade/repository/business/SpitRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SpitRepository extends JpaRepository <Spit, Integer> {
+public interface SpitRepository extends JpaRepository <Spit, Long> {
 }
